backend/requirements.txt
# FastAPI + Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Document Parsing
PyMuPDF==1.24.10        # PDF parsing
python-docx==1.1.2      # DOCX parsing
chardet==5.2.0          # Encoding detection

# AI / RAG
sentence-transformers==3.1.1   # Local embeddings
chromadb==0.5.7                # Vector database
google-generativeai==0.8.1     # Gemini (free API)
# cross-encoder for reranking (uses sentence-transformers)

# Utilities
python-dotenv==1.0.1
pydantic[email]==2.9.2
pydantic-settings==2.5.2

# Testing
httpx==0.27.2